+++ ui/src/components/config-form.shared.ts
@@ -1,72 +1,16 @@
 import { isSensitiveConfigPath } from "../../../src/config/sensitive-paths.js";
 import type { ConfigUiHint, ConfigUiHints } from "../api/types.ts";
-// Control UI view renders config form.shared screen content.
 import { t } from "../i18n/index.ts";
+import { hintForPath, pathKey } from "../lib/config-form-utils.ts";
 
-export type JsonSchema = {
-  type?: string | string[];
-  title?: string;
-  description?: string;
-  tags?: string[];
-  "x-tags"?: string[];
-  properties?: Record<string, JsonSchema>;
-  required?: string[];
-  items?: JsonSchema | JsonSchema[];
-  additionalItems?: JsonSchema | boolean;
-  additionalProperties?: JsonSchema | boolean;
-  enum?: unknown[];
-  enumIncludesNull?: boolean;
-  const?: unknown;
-  default?: unknown;
-  minimum?: number;
-  maximum?: number;
-  exclusiveMinimum?: number;
-  exclusiveMaximum?: number;
-  multipleOf?: number;
-  minLength?: number;
-  maxLength?: number;
-  pattern?: string;
-  minItems?: number;
-  maxItems?: number;
-  uniqueItems?: boolean;
-  anyOf?: JsonSchema[];
-  oneOf?: JsonSchema[];
-  allOf?: JsonSchema[];
-  nullable?: boolean;
-};
-
-export function schemaType(schema: JsonSchema): string | undefined {
-  if (!schema) {
-    return undefined;
-  }
-  if (Array.isArray(schema.type)) {
-    return schema.type.find((type) => type !== "null") ?? schema.type[0];
-  }
-  return schema.type;
-}
-
-export function schemaMayAcceptString(schema: JsonSchema): boolean {
-  const declaredTypes = Array.isArray(schema.type) ? schema.type : schema.type ? [schema.type] : [];
-  if (declaredTypes.length > 0 && !declaredTypes.includes("string")) {
-    return false;
-  }
-  if (schema.const !== undefined && typeof schema.const !== "string") {
-    return false;
-  }
-  if (schema.enum && !schema.enum.some((entry) => typeof entry === "string")) {
-    return false;
-  }
-  if (schema.allOf && !schema.allOf.every(schemaMayAcceptString)) {
-    return false;
-  }
-  if (schema.anyOf && !schema.anyOf.some(schemaMayAcceptString)) {
-    return false;
-  }
-  if (schema.oneOf && !schema.oneOf.some(schemaMayAcceptString)) {
-    return false;
-  }
-  return true;
-}
+export {
+  hintForPath,
+  humanize,
+  pathKey,
+  schemaMayAcceptString,
+  schemaType,
+  type JsonSchema,
+} from "../lib/config-form-utils.ts";
 
 export function configFieldId(path: Array<string | number>, suffix: string): string {
   const key =
@@ -84,47 +28,6 @@ export function configFieldId(path: Array<string | number>, suffix: string): str
           })
           .join("_");
   return `config-field-${key}-${suffix}`;
-}
-
-export function pathKey(path: Array<string | number>): string {
-  return path.filter((segment) => typeof segment === "string").join(".");
-}
-
-export function hintForPath(path: Array<string | number>, hints: ConfigUiHints) {
-  const key = pathKey(path);
-  const direct = hints[key];
-  if (direct) {
-    return direct;
-  }
-  const segments = path.map(String);
-  for (const [hintKey, hint] of Object.entries(hints)) {
-    if (!hintKey.includes("*")) {
-      continue;
-    }
-    const hintSegments = hintKey.split(".");
-    if (hintSegments.length !== segments.length) {
-      continue;
-    }
-    let match = true;
-    for (let i = 0; i < segments.length; i += 1) {
-      if (hintSegments[i] !== "*" && hintSegments[i] !== segments[i]) {
-        match = false;
-        break;
-      }
-    }
-    if (match) {
-      return hint;
-    }
-  }
-  return undefined;
-}
-
-export function humanize(raw: string) {
-  return raw
-    .replace(/_/g, " ")
-    .replace(/([a-z0-9])([A-Z])/g, "$1 $2")
-    .replace(/\s+/g, " ")
-    .replace(/^./, (m) => m.toUpperCase());
 }
 
 const ENV_VAR_PLACEHOLDER_PATTERN = /^\$\{[^}]*\}$/;
